working assisted soft clipper wdl
