+++ docker/lr-softclipper/Makefile
@@ -1,4 +1,4 @@
-VERSION = 0.2.0
+VERSION = 0.5.0
 TAG1 = quay.io/broad-long-read-pipelines/lr-softclipper:$(VERSION)
 TAG2 = quay.io/broad-long-read-pipelines/lr-softclipper:latest
 TAG3 = us.gcr.io/broad-dsp-lrma/lr-softclipper:$(VERSION)
@@ -8,10 +8,8 @@ all: build push
 
 build:
 	cp ../../scripts/soft_clipper.py .
-	cp ../../scripts/soft_clipper_aided.py .
-	docker build -t $(TAG1) -t $(TAG2) -t $(TAG3) -t $(TAG4) .
+	docker build --no-cache -t $(TAG1) -t $(TAG2) -t $(TAG3) -t $(TAG4) .
 	rm soft_clipper.py
-	rm soft_clipper_aided.py
 
 push:
 	docker push $(TAG1)
