+++ wdl/AssistedSoftClipperRunner.wdl
@@ -3,31 +3,31 @@ version 1.0
 import "tasks/SoftClipper.wdl" as SoftClipper
 import "tasks/Finalize.wdl" as FF
 
-workflow SoftClipperAidedRunner {
+workflow AssistedSoftClipperRunner {
     input {
         String reads_fastq
         File reference_fasta
         File aid_reference_fasta
         Int rounds
         Int clipping_threshold
-        Boolean keep_all_ref_conflicts
+        Int ref_conflict_threshold
 
         File output_dir
     }
 
-    call SoftClipper.SplitSoftClippedReadsAided {
+    call SoftClipper.SplitSoftClippedReadsAssisted {
         input:
             reads_fastq = reads_fastq,
             reference_fasta = reference_fasta,
             aid_reference_fasta = aid_reference_fasta,
             rounds = rounds,
             clipping_threshold = clipping_threshold,
-            keep_all_ref_conflicts = keep_all_ref_conflicts
+            ref_conflict_threshold = ref_conflict_threshold,
     }
 
     call FF.FinalizeToDir {
         input:
-            files = SplitSoftClippedReadsAided.split_reads,
+            files = flatten([SplitSoftClippedReadsAssisted.split_reads, SplitSoftClippedReadsAssisted.conflicting_alignments]),
             outdir = output_dir
     }
 }
